refactored to float
changed double to float, set workgroupsize  == max num particles

+++ Galaxy.h
@@ -5,6 +5,9 @@
 #include <iostream>
 #include "Tree.h"
 #include <Windows.h>
+#include <ctime>
+#include <thread>
+#include "OpenCLHelper.h"
 
 
 
@@ -12,30 +15,30 @@ class Galaxy
 {
 public:
 	Galaxy();
-	Galaxy(double x, double y, double _centerMass, int _NUM_P,
-		double vel_x, double vel_y, double radius);
+	Galaxy(float x, float y, float _centerMass, int _NUM_P,
+		float vel_x, float vel_y, float radius);
 	~Galaxy();
 	void displayParticles(std::vector<ParticleData*> arr, GLFWwindow* window);
 	void displayParticles2(std::vector<ParticleData*> arr1,
 		std::vector<ParticleData*> arr2, GLFWwindow* window);
 
-	void recursiveBoxes(QuadNode& qt, double factor);
+	void recursiveBoxes(QuadNode& qt, float factor);
 	void displayQuadrant(QuadNode& quad);
 	void displayQuadrant(QuadNode& quad, QuadNode& second);
 
 	int running_display();
 	int two_running_display(Galaxy& second);
 
-	void add_galaxy(Galaxy& galaxy, double vel_x, double vel_y);
-	double clockToMilliseconds(clock_t ticks);
+	void add_galaxy(Galaxy& galaxy, float vel_x, float vel_y);
+	float clockToMilliseconds(clock_t ticks);
 	static std::vector<ParticleData*> renegades;
 
 private:
 	ParticleData particle; 
 	std::vector<ParticleData*> allParticles;
 	QuadNode * root;
-	double x;
-	double y;
+	float x;
+	float y;
 	int NUMBER_PARTICLES;
 	Vector2D* max;
 	Vector2D* min;
