+++ ParticleData.h
@@ -18,25 +18,25 @@ struct ParticleData
 {
 public:
 	ParticleData();
-	ParticleData(double _x, double _y, double mstate);
-	ParticleData(double _x, double _y, double mstate, double _vx, double _vy);
+	ParticleData(float _x, float _y, float mstate);
+	ParticleData(float _x, float _y, float mstate, float _vx, float _vy);
 
 	~ParticleData() = default;
 	void reset();
 	bool isNull() const;
 	void printParticle();
-	std::vector<ParticleData*> generateParticles(double a, 
-		double b,int n, double R, double _centerMass, double vel_x, double vel_y);
+	std::vector<ParticleData*> generateParticles(float a, 
+		float b,int n, float R, float _centerMass, float vel_x, float vel_y);
 	void calcDistance(Vector2D force);
 
-	void createParticle(double a, double b, double radius, 
+	void createParticle(float a, float b, float radius, 
 		int start, int end, std::vector<ParticleData*> &arr,
-		double _centerMass, double vel_x, double vel_y);
+		float _centerMass, float vel_x, float vel_y);
 
 	//__global__ void calcDistance_CUDA(Vector2D* forces, Vector2D* particles, int count);
 
 
 	Vector2D* xy;
-	double mState;
+	float mState;
 };
